Determing the Winner & Final Touches
Here I have incorporated the criteria which determines the winner using a series of nested if statements. I've outputed the winner with a brief farewell message. I've also went back and touched up a few earlier components: primarily, limited the errors on the readiness check to a grand total of five.